+++ sk_kernel/mask.cuh
@@ -44,4 +44,4 @@ __device__ float d_V_func(float);
  * @param mu_min Minimum for mu to be considered good
  * @param N_good_min Minimum number of good feeds to continue
 */
-__global__ void mask(uint32_t *, uint32_t *, float *, float *, size_t, size_t, size_t, size_t, float, float, float, float*, float*, float*);
+__global__ void mask(uint32_t *, uint32_t *, uint *, uint *, size_t, size_t, size_t, size_t, float, float, float, float*, float*, float*);
